% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinkpatheR.R
\name{pinkpatheR}
\alias{pinkpatheR}
\title{A function for soft-coded, OS agnostic home directory absolut path retrieval}
\usage{
pinkpatheR()
}
\description{
This function allows you to have user specific absolut paths in scripts.
}
\examples{
homedir = pinkpatheR::pinkpatheR()
}
\keyword{pinkpather}
